+++ relecov_dashboard/templates/relecov_dashboard/variantsIndex.html
@@ -30,7 +30,7 @@
               <h5 class="card-title">Lineages</h5><!--Hackaton Graphs-->
               <p class="card-text">
                 <ul>
-                  <li><a href="#">Graph by time</a></li>
+                  <li><a href="/dashboard/variants/lineageVariationOverTime">Graph by time</a></li>
                   <li><a href="/dashboard/variants/geoJSON">Graph by Autonomous Community</a></li>
                   <li><a href="#">Table to filter and download</a></li>
                 </ul>
